+++ botrunner/internal/game/game_info.go
@@ -8,9 +8,9 @@ type GameInfo struct {
 	GameCode              string
 	GameType              string
 	Title                 string
-	SmallBlind            float32
-	BigBlind              float32
-	StraddleBet           float32
+	SmallBlind            int64
+	BigBlind              int64
+	StraddleBet           int64
 	UtgStraddleAllowed    bool
 	ButtonStraddleAllowed bool
 	MinPlayers            uint32
@@ -23,9 +23,9 @@ type GameInfo struct {
 	SitInApproval         bool
 	MaxWaitList           uint32
 	RakePercentage        float32
-	RakeCap               float32
-	BuyInMin              float32
-	BuyInMax              float32
+	RakeCap               int64
+	BuyInMin              int64
+	BuyInMax              int64
 	ActionTime            uint32
 	MuckLosingHand        bool
 	RunItTwiceAllowed     bool
@@ -56,32 +56,32 @@ type GameInfo struct {
 
 // SeatInfo is the info about a player sitting in a game.
 type SeatInfo struct {
-	SeatNo     uint32  `json:"seatNo"`
-	PlayerUUID string  `json:"playerUuid"`
-	PlayerId   uint64  `json:"playerId"`
-	Name       string  `json:"name"`
-	BuyIn      float32 `json:"buyIn"`
-	Stack      float32 `json:"stack"`
-	IsBot      bool    `json:"isBot"`
-	Status     string  `json:"status"`
+	SeatNo     uint32 `json:"seatNo"`
+	PlayerUUID string `json:"playerUuid"`
+	PlayerId   uint64 `json:"playerId"`
+	Name       string `json:"name"`
+	BuyIn      int64  `json:"buyIn"`
+	Stack      int64  `json:"stack"`
+	IsBot      bool   `json:"isBot"`
+	Status     string `json:"status"`
 }
 
 // GameCreateOpt contains parameters for creating a new game.
 type GameCreateOpt struct {
 	Title              string
 	GameType           string
-	SmallBlind         float32
-	BigBlind           float32
+	SmallBlind         int64
+	BigBlind           int64
 	UtgStraddleAllowed bool
-	StraddleBet        float32
+	StraddleBet        int64
 	MinPlayers         int
 	MaxPlayers         int
 	GameLength         int
 	BuyInApproval      bool
 	RakePercentage     float32
-	RakeCap            float32
-	BuyInMin           float32
-	BuyInMax           float32
+	RakeCap            int64
+	BuyInMin           int64
+	BuyInMax           int64
 	ActionTime         int
 	RewardIds          []uint32
 	RunItTwiceAllowed  bool
